rws_py/requirements.txt
urllib3 >= 1.15.1
websocket-client>=1.4.2
requests>=2.28.1
setuptools >= 21.0.0
numpy >= 1.19.2
xmltodict >= 0.13.0
